Implement basic 8points based initialisation
For the moment only for two cameras, PnP should still be added for the
other cameras and Triangulation should be used for the missing points.

Note: The model could also be extended to estimate the depth of the
initial points in the initial camera frame using more than one
additional camera.