Make the ModularSBASolver more modular
This version tries to ignore which kind of blocks are created, instead
being concerned only with positions, poses, trajectories and projector.

As much work as possible is now delegated to the modules.

This also make it possible to be as generic as possible with the
correspondencesset datablock, now we can have a dedicated module for
these, instead of delegating the use of correspondences to all the other
modules for individual datablocks.