Better logging capabilities for the modular sba solver
The loggers can now use functions that are not managed by the ceres
problem.

This is usefull for example when one wants to log the error without
considering the covariance matrix, so the logging error function will
have unit stiffness while the one used by ceres will get the proper
stiffness matrix.