Add debugging tools for cost function diagnosis
Add a class in modular sba solver cost functions can inherist from,
which allow to attach references to a cost function. This is not very
usefull for final implementation, but can be used for debugging
purposes, when cost functions needs to be inspected with a debugger, or
when they need to print some sort of self diagnosis.